+++ variants/tests/test_forms.py
@@ -18,7 +18,7 @@ class TestFilterForm(TestFormBase):
 
     def test_submit_defaults(self):
         form_data = vars(FormDataFactory(names=self.variant_set.case.get_members()))
-        form = FilterForm(form_data, case=self.variant_set.case)
+        form = FilterForm(form_data, case=self.variant_set.case, user=0)
         self.assertTrue(form.is_valid())
 
     def test_genomic_region_range(self):
@@ -27,7 +27,7 @@ class TestFilterForm(TestFormBase):
                 names=self.variant_set.case.get_members(), genomic_region="19:1,000,000-2,000,000"
             )
         )
-        form = FilterForm(form_data, case=self.variant_set.case)
+        form = FilterForm(form_data, case=self.variant_set.case, user=0)
         self.assertTrue(form.is_valid())
         self.assertEqual(form.cleaned_data["genomic_region"], [("19", 1000000, 2000000)])
 
@@ -35,6 +35,6 @@ class TestFilterForm(TestFormBase):
         form_data = vars(
             FormDataFactory(names=self.variant_set.case.get_members(), genomic_region="X")
         )
-        form = FilterForm(form_data, case=self.variant_set.case)
+        form = FilterForm(form_data, case=self.variant_set.case, user=0)
         self.assertTrue(form.is_valid())
         self.assertEqual(form.cleaned_data["genomic_region"], [("X", None, None)])